+++ eslint.config.js
@@ -1,33 +1,38 @@
-import js from "@eslint/js";
-import globals from "globals";
-import tseslint from "typescript-eslint";
+import js from '@eslint/js';
+import globals from 'globals';
+import tseslint from 'typescript-eslint';
+import prettierPlugin from 'eslint-plugin-prettier';
+import prettierConfig from 'eslint-config-prettier';
 
 export default [
   js.configs.recommended,
   ...tseslint.configs.recommended,
+  prettierConfig,
   {
-    files: ["**/*.ts"],
+    files: ['**/*.ts'],
     languageOptions: {
       globals: {
         ...globals.node,
       },
       parser: tseslint.parser,
       parserOptions: {
-        project: "./tsconfig.json",
+        project: './tsconfig.json',
       },
     },
     plugins: {
-      "@typescript-eslint": tseslint.plugin,
+      '@typescript-eslint': tseslint.plugin,
+      prettier: prettierPlugin,
     },
     rules: {
       ...tseslint.configs.recommended.rules,
-      "no-const-assign": "warn",
-      "no-this-before-super": "warn",
-      "no-undef": "warn",
-      "no-unreachable": "warn",
-      "no-unused-vars": "warn",
-      "constructor-super": "warn",
-      "valid-typeof": "warn",
+      'prettier/prettier': 'error',
+      'no-const-assign': 'warn',
+      'no-this-before-super': 'warn',
+      'no-undef': 'warn',
+      'no-unreachable': 'warn',
+      'no-unused-vars': 'warn',
+      'constructor-super': 'warn',
+      'valid-typeof': 'warn',
     },
   },
 ];